Refactor client auth flow to redux action
