+++ client/app/actions/index.js
@@ -1,5 +1,7 @@
-import Expo from 'expo';
+import Expo, { AuthSession } from 'expo';
 import { times } from 'lodash';
+
+import { LI_APP_ID } from 'react-native-dotenv';
 
 import {
   FETCH_MATCHES,
@@ -9,6 +11,8 @@ import {
   LOGOUT,
   MODIFY_USER_PROFILE,
 } from './actionTypes';
+
+import { fetchIdToken } from '../firebase';
 
 const useMocks = true;
 
@@ -89,6 +93,32 @@ export const modifyUserProfile = userProfile => ({
   type: MODIFY_USER_PROFILE,
   userProfile
 });
+
+export const authenticateAndCreateProfile = () => (
+  async dispatch => {
+    const redirectUri = AuthSession.getRedirectUrl();
+    const result = await AuthSession.startAsync({
+      authUrl:
+        `https://www.linkedin.com/oauth/v2/authorization?response_type=code` +
+        `&client_id=${LI_APP_ID}` +
+        `&redirect_uri=${encodeURIComponent(redirectUri)}` +
+        `&state=meetover_testing`
+    });
+
+    if (result.type === 'success') {
+      const uri = `https://meetover.herokuapp.com/login/${result.params.code}` +
+        `?redirect_uri=${encodeURIComponent(redirectUri)}`;
+      const init = { method: 'POST' };
+
+      const response = await fetch(uri, init);
+      const { profile, token, firebaseCustomToken } = await response.json();
+      const firebaseIdToken = await fetchIdToken(firebaseCustomToken)
+        .catch(err => null);
+
+      dispatch(createProfile({ ...profile, token, firebaseCustomToken }));
+    }
+  }
+);
 
 export const saveProfileAndLoginAsync = userProfile => (
   dispatch => (
