+++ client/app/screens/LoginScreen.js
@@ -1,16 +1,12 @@
 import React from 'react';
 import { StyleSheet } from 'react-native';
-import { AuthSession } from 'expo';
 import { Button, View, Spinner } from 'native-base';
 
 import { connect } from 'react-redux';
 
-import { LI_APP_ID } from 'react-native-dotenv';
-
 import Colors from '../constants/Colors';
 import { PTSansText } from '../components/StyledText';
-import { createProfile, saveProfileAndLoginAsync } from '../actions';
-import { fetchIdToken } from '../firebase';
+import { authenticateAndCreateProfile } from '../actions';
 
 class LoginScreen extends React.Component {
   static navigationOptions = {
@@ -53,35 +49,14 @@ class LoginScreen extends React.Component {
   }
 
   _handleSignInPressAsync = async () => {
-    const redirectUri = AuthSession.getRedirectUrl();
-    const result = await AuthSession.startAsync({
-      authUrl:
-        `https://www.linkedin.com/oauth/v2/authorization?response_type=code` +
-        `&client_id=${LI_APP_ID}` +
-        `&redirect_uri=${encodeURIComponent(redirectUri)}` +
-        `&state=meetover_testing`
-    });
-
-    if (result.type === 'success') {
-      this.setState({ isLoading: true });
-
-      const uri = `https://meetover.herokuapp.com/login/${result.params.code}` +
-        `?redirect_uri=${encodeURIComponent(redirectUri)}`;
-      const init = { method: 'POST' };
-
-      const response = await fetch(uri, init);
-      const { profile, token, firebaseCustomToken } = await response.json();
-      const firebaseIdToken = await fetchIdToken(firebaseCustomToken)
-        .catch(err => null);
-
-      this.props.createProfile({ ...profile, token, firebaseCustomToken });
-      this.setState({ isLoading: false });
-    }
+    this.setState({ isLoading: true });
+    await this.props.authenticateAndCreateProfile();
+    this.setState({ isLoading: false });
   }
 }
 
 const mapDispatchToProps = {
-  createProfile
+  authenticateAndCreateProfile
 };
 
 export default connect(
